Move modal hook to hooks folder
