+++ site/gatsby-site/src/pages/apps/classifications.js
@@ -12,7 +12,7 @@ import Link from 'components/ui/Link';
 import { faExpandAlt } from '@fortawesome/free-solid-svg-icons';
 import { faEdit } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { useModal, CustomModal } from 'components/useModal';
+import { useModal, CustomModal } from 'hooks/useModal';
 import { useUserContext } from 'contexts/userContext';
 import DateRangePicker from 'react-bootstrap-daterangepicker';
 import { format } from 'date-fns';
