+++ site/gatsby-site/src/pages/apps/discover.js
@@ -4,7 +4,7 @@ import algoliasearch from 'algoliasearch/lite';
 import { InstantSearch } from 'react-instantsearch-dom';
 import LayoutHideSidebar from 'components/LayoutHideSidebar';
 import Helmet from 'react-helmet';
-import { useModal, CustomModal } from 'components/useModal';
+import { useModal, CustomModal } from 'hooks/useModal';
 
 import config from '../../../config';
 import Hits from 'components/discover/Hits';
